Try splitting npz stuff out - Refactoring (WIP)
